Add language validation and image type validation attributes
